--- packages/base/src/formbuilder/objectform/process/index.ts
@@ -1,1 +1,0 @@
-export * from './dissolveProcessForm';
